# contracts/v1/openapi.yaml
openapi: 3.1.0

info:
  title: Emergency Scenario State API
  version: 1.1.0
  description: >
    Shared scenario-state contract used by the control-room
    and steward clients.

servers:
  - url: http://localhost:4010

paths:
  /api/v1/scenario-state:
    get:
      summary: Get current scenario state
      responses:
        "200":
          description: Current scenario state
          content:
            application/json:
              schema:
                $ref: "./scenario-state.schema.json"

  /api/v1/scenarios:
    get:
      summary: Get all scenarios
      responses:
        "200":
          description: List of scenarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "./scenario.schema.json"

  /api/v1/scenarios/{scenarioId}:
    get:
      summary: Get a single scenario
      parameters:
        - name: scenarioId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Scenario
          content:
            application/json:
              schema:
                $ref: "./scenario.schema.json"
        "404":
          description: Scenario not found

  /api/v1/metro-lines/{lineId}:
    get:
      summary: Get metro line by ID
      parameters:
        - name: lineId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Metro line details
          content:
            application/json:
              schema:
                $ref: "./metro-line.schema.json"

  /api/v1/position-reports:
    post:
      summary: Report a steward position
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./position-report.schema.json"
            example:
              staff: "steward-002"
              station: "Station B"
              reportedAt: "2026-09-23T08:10:00Z"

      responses:
        "202":
          description: Position report accepted
          content:
            application/json:
              schema:
                type: object
                required:
                  - accepted
                properties:
                  accepted:
                    type: boolean
              example:
                accepted: true

  /api/v1/scenarios/{scenarioId}/activate:
    post:
      summary: Activates a scenario with given id
      parameters:
        - name: scenarioId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./activation-request.schema.json"
            example:
              actor: "operator-017"
      responses:
        "201":
          description: Activation successful
          content:
            application/json:
              schema:
                $ref: "./scenario-state.schema.json"
        "400":
          description: Invalid request body, e.g. missing actor.
          content:
            application/json:
              schema:
                $ref: "./error.schema.json"
              example:
                code: "INVALID_REQUEST"
                message: "Request body must have required property 'actor'"
        "404":
          description: Scenario not found
          content:
            application/json:
              schema:
                $ref: "./error.schema.json"
              example:
                code: "SCENARIO_NOT_FOUND"
                message: "No scenario with id 'S1' exists"
        "409":
          description: >
            Activation refused because this same scenario is already
            active. Other active scenarios, including scenarios on the
            same line, never block an activation and are never stood
            down by it.
          content:
            application/json:
              schema:
                $ref: "./error.schema.json"
              example:
                code: "SCENARIO_ALREADY_ACTIVE"
                message: "Scenario 'S1' is already active"